+++ mybatis-plus/src/test/java/com/baomidou/mybatisplus/test/h2/entity/H2Student.java
@@ -13,15 +13,15 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.baomidou.mybatisplus.test.h2.entity.persistent;
+package com.baomidou.mybatisplus.test.h2.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.extension.activerecord.Model;
-import com.baomidou.mybatisplus.test.h2.entity.enums.GenderEnum;
-import com.baomidou.mybatisplus.test.h2.entity.enums.GradeEnum;
 
+import com.baomidou.mybatisplus.test.h2.enums.GenderEnum;
+import com.baomidou.mybatisplus.test.h2.enums.GradeEnum;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
