+++ mybatis-plus/src/test/java/com/baomidou/mybatisplus/test/h2/enums/GradeEnum.java
@@ -13,13 +13,37 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.baomidou.mybatisplus.test.base.mapper.pg;
+package com.baomidou.mybatisplus.test.h2.enums;
 
-import com.baomidou.mybatisplus.test.base.entity.pg.PgData;
-import com.baomidou.mybatisplus.test.base.mapper.MyBaseMapper;
+import com.baomidou.mybatisplus.annotation.EnumValue;
 
 /**
- * @author miemie
- * @since 2018-08-06
+ * <p></p>
+ *
+ * @author yuxiaobin
+ * @date 2018/8/30
  */
-public interface PgDataMapper extends MyBaseMapper<PgData> {}
+public enum GradeEnum {
+
+    PRIMARY(1, "小学"),
+    SECONDARY(2, "中学"),
+    HIGH(3, "高中");
+
+    GradeEnum(int code, String descp) {
+        this.code = code;
+        this.descp = descp;
+    }
+
+    @EnumValue
+    private final int code;
+
+    private final String descp;
+
+    public int getCode() {
+        return code;
+    }
+
+    public String getDescp() {
+        return descp;
+    }
+}
