+++ mybatis-plus/src/test/java/com/baomidou/mybatisplus/test/h2/service/IH2StudentService.java
@@ -16,7 +16,7 @@
 package com.baomidou.mybatisplus.test.h2.service;
 
 import com.baomidou.mybatisplus.extension.service.IService;
-import com.baomidou.mybatisplus.test.h2.entity.persistent.H2Student;
+import com.baomidou.mybatisplus.test.h2.entity.H2Student;
 
 /**
  * Created by nieqiuqiu on 2018/9/6.
