+++ mybatis-plus/src/test/java/com/baomidou/mybatisplus/test/h2/service/IH2UserService.java
@@ -20,7 +20,7 @@ import java.util.Map;
 
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.service.IService;
-import com.baomidou.mybatisplus.test.h2.entity.persistent.H2User;
+import com.baomidou.mybatisplus.test.h2.entity.H2User;
 
 /**
  * Service层测试
