+++ mybatis-plus/src/test/resources/mybatis-config.xml
@@ -9,6 +9,6 @@
     </settings>
 
     <mappers>
-        <mapper class="com.baomidou.mybatisplus.test.h2.entity.mapper.H2UserMapper"/>
+        <mapper class="com.baomidou.mybatisplus.test.h2.mapper.H2UserMapper"/>
     </mappers>
 </configuration>
